Rachel's visuals
top50_size_reviews_plot.pdf:
The lower the size, the more spread out across the board review counts are for the top 50 apps (ordered by most reviews), whereas the apps with really high sizes tend to have less reviews. Higher the size of the app, more chance for failure/crashing --> less visibility among the public because nobody's going out of their way to rate or review them?

cumulative_size_reviews_plot.pdf:
found that if you were to put into consideration ALL the data points (not just the top 50 with the most review counts), it's consistent with what we found in top50_size_reviews_plot.pdf as well. More chance for visibility among the public. The most spread out tends to be towards the lower end of the spectrum of the size of the app. We can infer that a small size for an app works inversely against the rate of its review count. Probably because the smaller the size of the app is, the less likely it is for the app to crash. We can venture a guess in that the higher the size of the app, the more likely it is to crash simply due to the amount of functionalities it has. Again, less exposure/popularity among the public.

Last_Updated_Reviews_Plots.pdf:
Newest apps/apps that consistently make updates/improvements to their apps tend to have more review counts. From this, we can gather the idea that the more update you make on an app, the higher the chances are of the user actively engaging with the app enough to leave comments and ratings for it. Increases visibility in the market. In comparison to the cumulative_size_reviews_plot and top50_size_reviews_plot, this means that it isn't necessarily ADDING features to an app that makes it more visible; it's improving the quality of what it has already without kicking up the size of the app so much that it crashes for the user. 

update_size_plot.pdf:
There is an interesting correlation between latest update and size in that the more often a company updates its app, the more spread out its size is across the y-axis. If you look at the section of apps updated solely in 2010, it has a really small app size.